refactor + length validation
